added updater

+++ uninstall.php
@@ -7,21 +7,39 @@
  */
 
 if (!defined('WP_UNINSTALL_PLUGIN')) {
-  $custom_post_types = [];
 
+  $custom_post_types = get_option('nrd_untappd_importer_menu') ?: array();
 
-  foreach ($custom_post_types as $post_type) {
-    $posts = get_posts([
-      'post_type' => $post_type,
-      'numberposts' => -1,
-      'post_status' => 'any'
-    ]);
+  if(count($custom_post_types) > 0) {
+    // Delete all posts
+    foreach ($custom_post_types as $post_type) {
+      $posts = get_posts([
+        'post_type' => $post_type['post_type'],
+        'numberposts' => -1,
+        'post_status' => 'any'
+      ]);
 
-    foreach ($posts as $post) {
-      wp_delete_posts($post->ID, true);
+      foreach ($posts as $post) {
+        wp_delete_posts($post->ID, true);
+      }
+    }
+
+    // Delete all terms of associated taxonomies
+    $taxonomies = get_object_taxonomies($post_type);
+    foreach ($taxonomies as $taxonomy) {
+      $terms = get_terms(
+        array(
+          'taxonomy' => $taxonomy,
+          'hide_empty' => false,
+        )
+      );
+
+      foreach ($terms as $term) {
+        wp_delete_term($term->term_id, $taxonomy);
+      }
     }
   }
-
+  
   $options = [
     'nrd_untappd_importer_schedule',
     'nrd_untappd_importer_api_creds',
